+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# 0.9.5
+
+### Notes
+
+* Notifications now use ChangeDetectionStrategy.OnPush
+
+# 0.9.4
+
+### Features
+
+* Extended position to support "middle" and "center" as options
+
+# 0.9.0
+
+### Notes
+
+* Refactored build system to support AOT and Angular version 5.
+
+### Features
+
+* Title and content now support TemplateRef and HTML
+
 # 0.7.0
 
 ### Breaking Changes